requirements.txt
h5py
numpy
Pillow
scikit-image
scipy
tensorflow-gpu==1.10.0
keras==2.2.4